modified payment and inventory order service

+++ order-service/src/main/java/com/shopping/order/models/OrderItems.java
@@ -29,6 +29,6 @@ import java.math.BigDecimal;
         private BigDecimal price;
         private Integer quantity;
         @Column(name="sub_total")
-        private Double subtotal;
+        private BigDecimal subtotal;
 
 }
